Change.log
- Added resumedev.sh to make it easier to resume development work without needing to use a permanent
  instance of screen or having to resume from the same host-machine.

In radioextract.sh:
   - Implemented de-dispersed search phase of radiotrans workflow.

In utils.sh:
   - Added explicit status returns to delete_files() and transfer_files()
   - Changed to using silenced pushd and popd to change directories in transfer_files()
   - Silenced the 'ls' command in transfer_files().
   - Added check for source and destination directories existing in transfer_files().  Now returns
     status >0 (failure) if either directory does not exist.
   - Silenced 'mv' command in transfer_files().
   - Added check in delete_files() to only delete paths that are actually files.
   - Added user feedback to delete_files().

In radiorun_lwa.sh:
   - Significantly reduced the commandline options.  Having all these options that just replicated
     radioreduce.sh is unnecessary.  We just need a debug configuration and the regular run
     configuration that gets modified for each run we wish to perform.
   - Renamed CMD to CMD_REDUCE and CMD_OPTS to CMD_REDUCE_OPTS.  This is in anticipation of building
     more phases into the LWA workflow (it is possible I may do the RFI and bandpass filtering on LWA
     along with the data reduction).
   - Modified to creating working and results directory structures recursively using 'mkdir -p'
     command.

In .gitignore:
   - Added dv.py.orig, which is a copy of the original dv.py file.  This is just so I don't have to deal
     with dancing around it during commits.

In dv.py:
   - Removed txtsize and replaced with pulseID.
   - Added writing of de-dispersion and pulse search parameters to common parameters file.
   - Renamed main_radiotrans() to just main().
   - Added exit status after call to main().
   - Changed from having the bandpass specified from the commandline to being read from the common
     parameters file.
   - Expanded width of pulseID field in output to 8.

In waterfall.py:
   - Removed main_orig().  We no longer to need to keep this around because the original idea of
     switching between the two main functions has been rendered obsolete by the CLeague branch.
   - Renamed main_radiotrans() to main()
   - Changed from importing individual functions from apputils.py to importing the entirety of
     apputils.py and prepending the functions from that module with the module name.  This was done to
     make it more clear that the functions are coming from a module and not being defined within
     waterfall.py.
   - Added exit status after call to main().

In waterfallcombine.py:
   - Changed from importing individual functions from apputils.py to importing the entirety of
     apputils.py and prepending the functions from that module with the module name.  This was done to
     make it more clear that the functions are coming from a module and not being defined within
     waterfallcombine.py.
   - Removed main_orig() and renamed main_radiotrans() to just main().
   - Added exit status after call to main().

In watchwaterfall.py:
   - Changed from importing individual functions from apputils.py to importing the entirety of
     apputils.py and prepending the functions from that module with the module name.  This was done to
     make it more clear that the functions are coming from a module and not being defined within
     watchwaterfall.py.
   - Encapsulated the main routine of code into main().
   - Changed section 'RFI Bandpass' in common parameters file to 'Spectrogram Plot'.  Want to use 'RFI
     Bandpass' for the parameters of the actual RFI-bandpass filtration stage.

In bandpasscheck.py:
   - Changed from importing individual functions from apputils.py to importing the entirety of
     apputils.py and prepending the functions from that module with the module name.  This was done to
     make it more clear that the functions are coming from a module and not being defined within
     bandpasscheck.py.
   - Removed main_orig() and renamed main_radiotrans() to just main().
   - Added exit status after call to main().

+++ utils.sh
@@ -18,8 +18,14 @@ function delete_files()
 
    for curr in ${FILES[*]}
    do
-      rm -f "${curr}"
+      if [ -f "${curr}" ]; then
+         echo "Deleting ${curr}"
+         rm -f "${curr}"
+      else
+         echo "${curr} not found or is not a file"
+      fi
    done
+   return 0
 }
 # end delete_files()
 
@@ -52,27 +58,40 @@ function transfer_files()
       done
    fi
 
-   # Check if we have a list of files to transfer.
-   if [ -z "${FILES}" ]; then
-      FILES=$(basename "${SRC_DIR}/*")
-   else
-      # If we are not in the source directory, then we will need to make sure we have an explicit list
-      # of all the files to be moved.
-      if [[ $(pwd) != "${SRC_DIR}" ]]; then
-         CURR_DIR=$(pwd)
-         cd "${SRC_DIR}"
-         FILES=$(ls ${FILES[*]})
-         cd "${CURR_DIR}"
+   # Check that the source and destination directories exist before proceding further.
+   if [ -d "${SRC_DIR}" -a -d "${DEST_DIR}" ]; then
+      # Check if we have a list of files to transfer.
+      if [ -z "${FILES[*]}" ]; then
+         # If no particular files specified, then transfer everything in the source directory.
+         FILES=$(basename "${SRC_DIR}/*")
+      else
+         # If we are not in the source directory, then we will need to make sure we have an explicit list
+         # of all the files to be moved.
+         if [[ $(pwd) != "${SRC_DIR}" ]]; then
+            pushd "${SRC_DIR}" 1>/dev/null
+            FILES=$(ls ${FILES[*]} 2>/dev/null)
+            popd 1>/dev/null
+         fi
       fi
-   fi
 
-   if [[ "${SRC_DIR}" != "${DEST_DIR}" ]]; then
-      echo "Transferring files from ${SRC_DIR} to ${DEST_DIR}..."
-      for curr in ${FILES[*]}
-      do
-         echo "Moving ${curr}"
-         mv "${SRC_DIR}/${curr}" "${DEST_DIR}/${curr}"
-      done
+      if [[ "${SRC_DIR}" != "${DEST_DIR}" ]]; then
+         echo "Transferring files from ${SRC_DIR} to ${DEST_DIR}..."
+         for curr in ${FILES[*]}
+         do
+            echo "Moving ${curr}"
+            mv "${SRC_DIR}/${curr}" "${DEST_DIR}/${curr}" 2>/dev/null
+         done
+      fi
+   else
+      if [ ! -d "${SRC_DIR}" ]; then
+         echo "Source directory ${SRC_DIR} does not exist or is not a directory."
+      fi
+      if [ ! -d "${DEST_DIR}" ]; then
+         echo "Destination directory ${DEST_DIR} does not exist or is not a directory."
+      fi
+      return 1
    fi
+   
+   return 0
 }
 # end transfer_files()
